% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-buildkite.R
\name{BuildkiteReporter}
\alias{BuildkiteReporter}
\title{Test reporter: Buildkite format.}
\description{
This reporter is an extension of ParallelProgressReporter see
source https://github.com/r-lib/testthat/blob/main/R/reporter-progress.R#L386
The overriden functions here are modified versions of the functions
on the ParallelProgressReporter.
}
\details{
This reporter gives us
\itemize{
\item Regularly updates progress logs
\item Time for each test file to run
\item All errors and warnings
}

This overrides some functions so that
\itemize{
\item It is less chatty (updates every 1 second) so build log is not too long
\item There is no spinner (as we are not running interactively)
}

In the future we could extend this to
\itemize{
\item Have more foldable sections
\item Report time for every test within a file
}
}
\concept{reporters}
\section{Super classes}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{\link[testthat:ProgressReporter]{testthat::ProgressReporter}} -> \code{\link[testthat:ParallelProgressReporter]{testthat::ParallelProgressReporter}} -> \code{BuidlkiteReporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BuildkiteReporter$new()}}
\item \href{#method-update}{\code{BuildkiteReporter$update()}}
\item \href{#method-clone}{\code{BuildkiteReporter$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context">}\href{../../testthat/html/Reporter.html#method-.start_context}{\code{testthat::Reporter$.start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line">}\href{../../testthat/html/Reporter.html#method-cat_line}{\code{testthat::Reporter$cat_line()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight">}\href{../../testthat/html/Reporter.html#method-cat_tight}{\code{testthat::Reporter$cat_tight()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context_if_started">}\href{../../testthat/html/Reporter.html#method-end_context_if_started}{\code{testthat::Reporter$end_context_if_started()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_test">}\href{../../testthat/html/Reporter.html#method-end_test}{\code{testthat::Reporter$end_test()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="local_user_output">}\href{../../testthat/html/Reporter.html#method-local_user_output}{\code{testthat::Reporter$local_user_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule">}\href{../../testthat/html/Reporter.html#method-rule}{\code{testthat::Reporter$rule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_test">}\href{../../testthat/html/Reporter.html#method-start_test}{\code{testthat::Reporter$start_test()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="cr">}\href{../../testthat/html/ProgressReporter.html#method-cr}{\code{testthat::ProgressReporter$cr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="is_full">}\href{../../testthat/html/ProgressReporter.html#method-is_full}{\code{testthat::ProgressReporter$is_full()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="report_issues">}\href{../../testthat/html/ProgressReporter.html#method-report_issues}{\code{testthat::ProgressReporter$report_issues()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="should_update">}\href{../../testthat/html/ProgressReporter.html#method-should_update}{\code{testthat::ProgressReporter$should_update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="show_status">}\href{../../testthat/html/ProgressReporter.html#method-show_status}{\code{testthat::ProgressReporter$show_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ProgressReporter" data-id="start_reporter">}\href{../../testthat/html/ProgressReporter.html#method-start_reporter}{\code{testthat::ProgressReporter$start_reporter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ParallelProgressReporter" data-id="add_result">}\href{../../testthat/html/ParallelProgressReporter.html#method-add_result}{\code{testthat::ParallelProgressReporter$add_result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ParallelProgressReporter" data-id="end_context">}\href{../../testthat/html/ParallelProgressReporter.html#method-end_context}{\code{testthat::ParallelProgressReporter$end_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ParallelProgressReporter" data-id="end_file">}\href{../../testthat/html/ParallelProgressReporter.html#method-end_file}{\code{testthat::ParallelProgressReporter$end_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ParallelProgressReporter" data-id="end_reporter">}\href{../../testthat/html/ParallelProgressReporter.html#method-end_reporter}{\code{testthat::ParallelProgressReporter$end_reporter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ParallelProgressReporter" data-id="show_header">}\href{../../testthat/html/ParallelProgressReporter.html#method-show_header}{\code{testthat::ParallelProgressReporter$show_header()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ParallelProgressReporter" data-id="start_context">}\href{../../testthat/html/ParallelProgressReporter.html#method-start_context}{\code{testthat::ParallelProgressReporter$start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ParallelProgressReporter" data-id="start_file">}\href{../../testthat/html/ParallelProgressReporter.html#method-start_file}{\code{testthat::ParallelProgressReporter$start_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="ParallelProgressReporter" data-id="status_data">}\href{../../testthat/html/ParallelProgressReporter.html#method-status_data}{\code{testthat::ParallelProgressReporter$status_data()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildkiteReporter$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildkiteReporter$update(force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildkiteReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
